fix temporal_stride

+++ multimodal_ropes/configs/hope.py
@@ -11,8 +11,8 @@ class HopeConfig(MRopeConfig):
         dim: int,
         base: int = 10000,
         mrope_section: list[int] = [16, 24, 24],
-        temporal_stride: int = 2,
-        temporal_stride_lst: list[int] = [0.5, 0.75, 1.0, 1.25, 1.5],
+        temporal_stride: float = 2.0,
+        temporal_stride_lst: list[float] = [0.5, 0.75, 1.0, 1.25, 1.5],
         **kwargs,
     ):
         super().__init__(dim, base, mrope_section, temporal_stride, **kwargs)
