+++ multimodal_ropes/configs/videorope.py
@@ -11,7 +11,7 @@ class VideoRopeConfig(MRopeConfig):
         dim: int,
         base: int = 10000,
         mrope_section: list[int] = [16, 24, 24],
-        temporal_stride: int = 2,
+        temporal_stride: float = 2.0,
         **kwargs,
     ):
         super().__init__(dim, base, mrope_section, temporal_stride, **kwargs)
